Split up an integration test class.
