+++ subprojects/file-collections/src/integTest/groovy/org/gradle/api/internal/file/FileCollectionIntegrationTest.groovy
@@ -59,153 +59,6 @@ class FileCollectionIntegrationTest extends AbstractIntegrationSpec implements T
         outputContains("files = [file1, file2, file3]")
     }
 
-    def "finalized file collection resolves locations and ignores later changes to source paths"() {
-        buildFile << """
-            def files = objects.fileCollection()
-            Integer counter = 0
-            files.from { "a\${++counter}" }
-
-            def names = ['b', 'c']
-            files.from(names)
-
-            assert files.files as List == [file('a1'), file('b'), file('c')]
-
-            files.finalizeValue()
-
-            assert counter == 2
-
-            assert files.files as List == [file('a2'), file('b'), file('c')]
-
-            counter = 45
-            names.clear()
-
-            assert files.files as List == [file('a2'), file('b'), file('c')]
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "finalize on read file collection resolves locations and ignores later changes to source paths"() {
-        buildFile << """
-            def files = objects.fileCollection()
-            Integer counter = 0
-            files.from { "a\${++counter}" }
-
-            def names = ['b', 'c']
-            files.from(names)
-
-            assert files.files as List == [file('a1'), file('b'), file('c')]
-
-            files.finalizeValueOnRead()
-
-            assert counter == 1
-
-            assert files.files as List == [file('a2'), file('b'), file('c')]
-
-            counter = 45
-            names.clear()
-
-            assert files.files as List == [file('a2'), file('b'), file('c')]
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "finalized file collection ignores later changes to nested file collections"() {
-        buildFile << """
-            def nested = objects.fileCollection()
-            def name = 'a'
-            nested.from { name }
-
-            def names = ['b', 'c']
-            nested.from(names)
-
-            def files = objects.fileCollection()
-            files.from(nested)
-            files.finalizeValue()
-            name = 'ignore-me'
-            names.clear()
-            nested.from('ignore-me')
-
-            assert files.files as List == [file('a'), file('b'), file('c')]
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "finalized file collection still reflects changes to file system but not changes to locations"() {
-        buildFile << """
-            def files = objects.fileCollection()
-            def name = 'a'
-            files.from {
-                fileTree({ name }) {
-                    include '**/*.txt'
-                }
-            }
-
-            files.finalizeValue()
-            name = 'b'
-
-            assert files.files.empty
-
-            file('a').mkdirs()
-            def f1 = file('a/thing.txt')
-            f1.text = 'thing'
-
-            assert files.files as List == [f1]
-        """
-
-        expect:
-        succeeds()
-    }
-
-    def "cannot mutate finalized file collection"() {
-        buildFile << """
-            def files = objects.fileCollection()
-            files.finalizeValue()
-
-            task broken {
-                doLast {
-                    files.from('bad')
-                }
-            }
-        """
-
-        when:
-        fails('broken')
-
-        then:
-        failure.assertHasCause("The value for this file collection is final and cannot be changed.")
-    }
-
-    def "can disallow changes to file collection without finalizing value"() {
-        buildFile << """
-            def files = objects.fileCollection()
-            def name = 'other'
-            files.from { name }
-
-            def names = ['b', 'c']
-            files.from(names)
-
-            files.disallowChanges()
-            name = 'a'
-            names.clear()
-
-            assert files.files as List == [file('a')]
-
-            files.from('broken')
-        """
-
-        when:
-        fails('broken')
-
-        then:
-        failure.assertHasCause("The value for this file collection cannot be changed.")
-    }
-
     def "can view the elements of file collection as a Provider"() {
         buildFile << """
             def files = objects.fileCollection()
@@ -219,55 +72,6 @@ class FileCollectionIntegrationTest extends AbstractIntegrationSpec implements T
 
         expect:
         succeeds()
-    }
-
-    def "task @InputFiles file collection property is implicitly finalized and changes ignored when task starts execution"() {
-        taskTypeWithInputFileCollection()
-        buildFile << """
-            task merge(type: InputFilesTask) {
-                outFile = file("out.txt")
-                inFiles.from = "in.txt"
-                doFirst {
-                    inFiles.from("other.txt")
-                }
-            }
-"""
-        file("in.txt").text = "in"
-
-        when:
-        executer.expectDocumentedDeprecationWarning("Changing the value for a FileCollection with a final value has been deprecated. " +
-            "This will fail with an error in Gradle 7.0. " +
-            "See https://docs.gradle.org/current/userguide/lazy_configuration.html#unmodifiable_property for more details.")
-        run("merge")
-
-        then:
-        file("out.txt").text == "in"
-    }
-
-    def "task ad hoc input file collection property is implicitly finalized and changes ignored when task starts execution"() {
-        buildFile << """
-            def files = project.files()
-            def outFile = file("out.txt")
-            task show {
-                inputs.files files
-                outputs.file outFile
-                files.from("in.txt")
-                doFirst {
-                    files.from("other.txt")
-                    outFile.text = files.files*.name.join(',')
-                }
-            }
-"""
-        file("in.txt").text = "in"
-
-        when:
-        executer.expectDocumentedDeprecationWarning("Changing the value for a FileCollection with a final value has been deprecated. " +
-            "This will fail with an error in Gradle 7.0. " +
-            "See https://docs.gradle.org/current/userguide/lazy_configuration.html#unmodifiable_property for more details.")
-        run("show")
-
-        then:
-        file("out.txt").text == "in.txt"
     }
 
     @ToBeFixedForInstantExecution
